chore: strated css generator
