+++ src/themeNpm/ThemeProcessor.ts
@@ -73,8 +73,9 @@ export class ThemeProcessor implements ObserverableThemeProcessor {
     if (cacheMiss) {
       this.processThemeData()
     } else {
-      this._state = 'data-ready'
-      this.notifyAll()
+      /* this._state = 'data-ready'
+      this.notifyAll() */
+      this.processThemeData()
     }
   }
